% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_add_scenario_type}
\alias{bs_add_scenario_type}
\title{Add an auc:ScenarioType element to the xml_node provided.  The node should be an auc:Scenario.}
\usage{
bs_add_scenario_type(
  x,
  sc_type = c("Current Building", "Package of Measures"),
  measured = TRUE
)
}
\arguments{
\item{x}{An xml_node for an auc:Scenario}

\item{sc_type}{A character string indicating the 'type' of the scenario: "Current Building", "Package of Measures"}

\item{measured}{A boolean indicating if this represents measured or modeled data.}
}
\value{
x An xml_node for the same auc:Scenario, with addition of the Scenario typing information.
}
\description{
Add an auc:ScenarioType element to the xml_node provided.  The node should be an auc:Scenario.
}
