% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_gen_dm_nmecr}
\alias{bs_gen_dm_nmecr}
\title{Add inputs, parameters, and performance statistics to a auc:DerivedModel}
\usage{
bs_gen_dm_nmecr(
  nmecr_baseline_model,
  x,
  normalization_method = "Forecast",
  response_variable = "Electricity",
  response_variable_units = "kWh",
  response_variable_end_use = "All end uses",
  explanatory_variable_name = "Drybulb Temperature",
  explanatory_variable_units = "Fahrenheit, F"
)
}
\arguments{
\item{nmecr_baseline_model}{An output from one of the nmecr_model_with_** functions.}

\item{x}{An xml_node for an auc:DerivedModel}

\item{normalization_method}{A character string representing the normalization method to employ.  See enums in BSync schema (auc:DerivedModelInputs/auc:NormalizationMethod).  This is most likely "Forecast".}

\item{explanatory_variable_units}{}
}
\value{
x An xml_node
}
\description{
Add inputs, parameters, and performance statistics to a auc:DerivedModel
}
