% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_gen_root_doc}
\alias{bs_gen_root_doc}
\title{Stub out the root node for an auc:BuildingSync document}
\usage{
bs_gen_root_doc(
 
    raw_schema_location = "https://raw.githubusercontent.com/BuildingSync/schema/c620c7e58688698901edcb8560cd3e1b4b34d971/BuildingSync.xsd"
)
}
\arguments{
\item{raw_schema_location}{A character string pointing to a raw XSD document to be used in xsi:schemaLocation}
}
\value{
doc An xml_document
}
\description{
Stub out the root node for an auc:BuildingSync document
}
