% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_link_bldg}
\alias{bs_link_bldg}
\title{Link the node provided to an auc:Building with the specified id}
\usage{
bs_link_bldg(x, linked_building_id)
}
\arguments{
\item{x}{An xml_node for an auc:Scenario}

\item{linked_building_id}{A character string representing the ID of the auc:Building to which this scenario should be linked}
}
\value{
x An xml_node for the same auc:Scenario, with the addition of the linked building.
}
\description{
Link the node provided to an auc:Building with the specified id
}
