% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_stub_bldg}
\alias{bs_stub_bldg}
\title{Stub out an auc:Building given a bare bones BuildingSync xml doc.}
\usage{
bs_stub_bldg(doc, bldg_id = "Building-1")
}
\arguments{
\item{doc}{An xml_document, similar to what is output by gen_root_doc}

\item{bldg_id}{A character string representing the id of the building wished to be used.  This should not have any spaces, and should start with a letter.  It is also used as the auc:PremisesName}
}
\value{
doc An xml_document, now with the building added
}
\description{
Stub out an auc:Building given a bare bones BuildingSync xml doc.
}
