% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_stub_derived_model}
\alias{bs_stub_derived_model}
\title{Stub out a Derived Model within an auc:Scenario}
\usage{
bs_stub_derived_model(
  x,
  dm_id,
  dm_period = c("Baseline", "Reporting"),
  sc_type = c("Current Building", "Package of Measures")
)
}
\arguments{
\item{x}{An xml_node for an auc:Scenario}

\item{dm_id}{The desired ID for the new derived model to be added}

\item{dm_period}{A character string to define the model period: "Baseline", "Reporting"}

\item{sc_type}{A character string indicating the 'type' of the scenario: "Current Building", "Package of Measures"}
}
\value{
x An xml_node for the same auc:Scenario, with the addition of the derived model
}
\description{
Stub out a Derived Model within an auc:Scenario
}
