% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bs_stub_scenarios}
\alias{bs_stub_scenarios}
\title{Add a Baseline and Reporting scenario to the document.  The document should already have an auc:Building element created.
It assumes that the Baseline and Reporting scenario both utilize actual measured building data.}
\usage{
bs_stub_scenarios(
  doc,
  baseline_id = "Scenario-Baseline",
  reporting_id = "Scenario-Reporting",
  linked_building_id = "Building-1"
)
}
\arguments{
\item{doc}{An xml_document, similar to what is output by stub_bldg}

\item{bldg_id}{A character string representing the id of the building wished to be used.  This should not have any spaces, and should start with a letter.  It is also used as the auc:PremisesName}
}
\value{
doc An xml_document, with two additional scenarios stubbed out.
}
\description{
Add a Baseline and Reporting scenario to the document.  The document should already have an auc:Building element created.
It assumes that the Baseline and Reporting scenario both utilize actual measured building data.
}
