% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsync_utils.R
\name{bsync_gen_root_doc}
\alias{bsync_gen_root_doc}
\title{Stub out the root node for an auc:BuildingSync document}
\usage{
bsync_gen_root_doc(
 
    raw_schema_location = "https://raw.githubusercontent.com/BuildingSync/schema/83e345b6706dea0154cfec9a402d2637f31c570a/BuildingSync.xsd"
)
}
\arguments{
\item{raw_schema_location}{A character string pointing to a raw XSD document to be used in xsi:schemaLocation}
}
\value{
doc An xml_document
}
\description{
Stub out the root node for an auc:BuildingSync document
}
